Add advice that works (.config symlinking)
This involves using a script to set the KCONFIG_OVERWRITECONFIG
environment variable, which is read by the kernel's build system and
tells it to not do the thing that makes the symlink fail.

So now there is a bunch of extra complication to get that to happen, but
hey, it works, and once it's done, it should _stay_ done.

Also renamed `kernel-config` -> `k.config`. I like this name better, as
it is more congruent with the `.config` name used in the kernel's
directory. And unlike `.config`, it is visible by default. That's
important, otherwise I would have just named it `.config`.